backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# 数据库
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1

# AI/ML
langchain==0.1.0
langchain-community==0.0.13
chromadb==0.4.22
sentence-transformers==2.3.1

# RAG Enhancement
rank-bm25>=0.2.2

# 工具
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1
python-multipart==0.0.6

# PDF处理
pypdf==4.0.1
python-docx==1.1.0

# Ollama客户端
ollama==0.1.6
google-generativeai>=0.3.0

# Azure Integration
openai>=1.0.0
azure-search-documents>=11.4.0
azure-ai-textanalytics>=5.3.0
azure-identity>=1.15.0
pytest
pytest-asyncio
httpx
